refactor dispute decode
